Add files via upload
Fixing batch size to run the example. /Easy-Beginner/. The GPU was running out of memory while running the code so the batch size had to be lowered from 128 to 10 in order to be able to run. This should be flexible depends on the GPU but my 2080ti can handle batch size 10 comfortably. the rest of the error messages can be ignored, such as delay kernel timed out:... . I hope this helps.