Add files via upload
Fixing batch size within the code to fix the running out of memory error message. /Easy/
In order to run the code on the GPU after setting up the ubuntu server from the step by step guide i had to lower the batch size to match the hardware limitation.
It is very easy but i thought i upload code to help to get started with machine learning. 
I simply changed the batch size from 128 to 10. The other errors can be ignored and the code should run smoothly with the right batch size. But a fresh start of the system and a fresh run should free up more GPU memory so the batch size can be maximized. This is just for self interest of course and running test on the GPU to see how it performs.