Add files via upload
$	pip install keras_hub

this will downgrade tensorflow 2.20 to 2.19. you need to uninstall keras_hub and reinstall tensorflow with cuda for the gpu after this example. 

$	pip uninstall keras-hub
$	pip uninstall tensorflow[and-cuda]
$	pip install tensorflow[and-cuda]
$      pip install --upgrade protobuf

So you reverted back from tf 2.19 to 2.20.

ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
tensorflow-text 2.19.0 requires tensorflow<2.20,>=2.19.0, but you have tensorflow 2.20.0 which is incompatible.

You can ignore that. 

ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
tensorflow-metadata 1.17.2 requires protobuf<4.22,>=4.21.6; python_version < "3.11", but you have protobuf 6.33.0 which is incompatible.
tensorflow-text 2.19.0 requires tensorflow<2.20,>=2.19.0, but you have tensorflow 2.20.0 which is incompatible.

You can ignore that as the system has multiple versions of cuda.