Add files via upload
Loaded runtime CuDNN library: 9.3.0 but source was compiled with: 9.8.0.  CuDNN library needs to have matching major version and equal or higher minor version. If using a binary install, upgrade your CuDNN library.  If building from sources, make sure the library loaded at runtime is compatible with the version specified during compile configuration.
Solution is to run it on the CPU for now:
A new virtualenv:

$	source bin/activate
$	pip install tensorflow[and-cuda]
$	pip install --upgrade pip	/if promted
$	pip install jupyterlab
$	pip install --upgrade keras

then:

$	pip list | grep nvidia

nvidia-cublas-cu12        12.9.1.4
nvidia-cuda-cupti-cu12    12.9.79
nvidia-cuda-nvcc-cu12     12.9.86
nvidia-cuda-nvrtc-cu12    12.9.86
nvidia-cuda-runtime-cu12  12.9.79
nvidia-cudnn-cu12         9.14.0.64
nvidia-cufft-cu12         11.4.1.4
nvidia-curand-cu12        10.3.10.19
nvidia-cusolver-cu12      11.7.5.82
nvidia-cusparse-cu12      12.5.10.65
nvidia-nccl-cu12          2.28.7
nvidia-nvjitlink-cu12     12.9.86
nvidia-nvshmem-cu12       3.4.5

$	pip install -U "jax[cuda13]"

then unsetting the LD_LIBRARY_PATH from the terminal:

$	unset LD_LIBRARY_PATH

An NVIDIA GPU may be present on this machine, but a CUDA-enabled jaxlib is not installed. Falling back to cpu.

It is for educational and for testing purposes only as it is an example and not a project. I suggest to stick to the support versions to have the gpu handle deep learning instead of the cpu. Tutorial can be found how to set up gpu support with the latest supported packages.